requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
iniconfig==1.1.1
lxml==4.8.0
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
scapy==2.4.5
soupsieve==2.3.2.post1
tomli==2.0.1
